+++ src/index.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import Home from './router'
+import './index.css'
+import Home from './pages/Home'
 
 ReactDOM.render(<Home />, document.getElementById('root'))
