+++ src/pages/App/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import ItemPost from '../ItemPost'
+import App from './index'
 
 it('renders without crashing', () => {
-  shallow(<ItemPost />)
+  shallow(<App />)
 })
