+++ src/pages/App/index.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import logo from './logo.svg'
-import './css/App.css'
+import logo from '../../logo.svg'
+import './App.css'
 
-function App() {
+export default function App() {
   return (
     <div className="App">
       <header className="App-header">
@@ -22,5 +22,3 @@ function App() {
     </div>
   )
 }
-
-export default App
