+++ src/pages/Home/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import ItemPost from './ItemPost'
-import App from './App'
+import ItemPost from '../ItemPost'
+import App from '../App'
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom'
 
 export default function Home() {
