+++ src/pages/ItemPost/ItemPost.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import App from '../App'
+import ItemPost from './index'
 
 it('renders without crashing', () => {
-  shallow(<App />)
+  shallow(<ItemPost />)
 })
